Hold the K shift in the cache's dtype, and subtract it in one kernel
The shift was kept in float32 against a bfloat16 cache, so every write
upcast, subtracted and downcast again. Three launches per layer, ninety-six
per step, which is 12.7% of an eager decode step and three times the memory
traffic the subtraction needs.

The float32 was defensive and measures as unnecessary. Softmax is invariant
to any constant removed from every key, however coarse that constant is, so
rounding it cannot make the result wrong -- only less effective at shrinking
the block scale. Against 11,578 captured post-RoPE K tokens across six
layers, the bfloat16 constant removes 55.80% of the NVFP4 error where the
float32 one removes 55.75%. The constant is rounded to 2e-3 relative and then
quantized to four bits, which is where that difference goes.

Measured on Llama-3.1-8B, batch 8: eager falls from +12.66% to +5.01% of a
decode step, and with CUDA graphs -- which the engine should be using anyway,
being 3.6x faster -- from +0.46% to +0.09%.

Loading moves below the tail allocation so the dtype comes from the buffer
being written rather than from a constant, and two tests pin the change: one
that the runtime's shift and its tail agree on dtype, one that a bfloat16
constant leaves attention exactly where it was.

+++ src/nvfp4_vllm/impl.py
@@ -147,7 +147,9 @@ class NVFP4Impl(FlashAttentionImpl):
             # Only what the cache holds is shifted, and it is shifted here
             # rather than in the quantizer so that the BF16 tail and the pages
             # promoted out of it agree with the pages written directly.
-            key = torch.sub(key.float(), shift[self.layer_index]).to(key.dtype)
+            # One kernel: the shift already carries K's dtype, so there is no
+            # upcast to undo afterwards.
+            key = torch.sub(key, shift[self.layer_index])
 
         key_pages_fp4, key_scales, value_pages_fp4, value_scales = (
             self.runtime.views(self.layer_index, kv_cache)
